+++ lua/vv-utils/fs/init.lua
@@ -1,6 +1,7 @@
 -- 文件系统公共入口
 
 local buffer = require('vv-utils.fs.buffer')
+local dir_info = require('vv-utils.fs.dir_info')
 local file_info = require('vv-utils.fs.file_info')
 local file_info_highlight = require('vv-utils.fs.file_info_highlight')
 local io = require('vv-utils.fs.io')
@@ -27,7 +28,12 @@ return {
   inspect_file = file_info.inspect,
   is_binary = file_info.is_binary,
   file_info_lines = file_info.lines,
+  format_size = file_info.format_size,
   highlight_file_info = file_info_highlight.apply,
+
+  inspect_dir = dir_info.shallow,
+  scan_dir = dir_info.scan,
+  dir_info_lines = dir_info.lines,
 
   sync_buffers = buffer.sync_buffers,
   new_transaction = transaction.new,